+++ src/readstat/spss/readstat_por_read.c
@@ -25,7 +25,7 @@
 #define POR_LABEL_NAME_PREFIX   "labels"
 
 #define MAX_FORMAT_TYPE       120
-#define MAX_FORMAT_WIDTH      255
+#define MAX_FORMAT_WIDTH    20000
 #define MAX_FORMAT_DECIMALS   100
 #define MAX_STRING_LENGTH   20000
 
@@ -511,8 +511,8 @@ static readstat_error_t read_variable_label_record(por_ctx_t *ctx) {
         goto cleanup;
     }
 
-    varinfo->label = realloc(varinfo->label, strlen(string) + 1);
-    strcpy(varinfo->label, string);
+    varinfo->label = realloc(varinfo->label, 4*strlen(string) + 1);
+    retval = readstat_convert(varinfo->label, 4*strlen(string) + 1, string, strlen(string), ctx->converter);
 
 cleanup:
     return retval;
@@ -697,7 +697,7 @@ readstat_error_t handle_variables(por_ctx_t *ctx) {
         spss_varinfo_t *info = &ctx->varinfo[i];
         info->index = i;
 
-        ctx->variables[i] = spss_init_variable_for_info(info, index_after_skipping);
+        ctx->variables[i] = spss_init_variable_for_info(info, index_after_skipping, ctx->converter);
 
         snprintf(label_name_buf, sizeof(label_name_buf), POR_LABEL_NAME_PREFIX "%d", info->labels_index);