+++ src/readstat/stata/readstat_dta_write.c
@@ -14,11 +14,12 @@
 
 #include "readstat_dta.h"
 
-#define DTA_DEFAULT_FORMAT_BYTE    "8.0g"
-#define DTA_DEFAULT_FORMAT_INT16   "8.0g"
-#define DTA_DEFAULT_FORMAT_INT32  "12.0g"
-#define DTA_DEFAULT_FORMAT_FLOAT   "9.0g"
-#define DTA_DEFAULT_FORMAT_DOUBLE "10.0g"
+#define DTA_DEFAULT_DISPLAY_WIDTH_BYTE   8
+#define DTA_DEFAULT_DISPLAY_WIDTH_INT16  8
+#define DTA_DEFAULT_DISPLAY_WIDTH_INT32  12
+#define DTA_DEFAULT_DISPLAY_WIDTH_FLOAT  9
+#define DTA_DEFAULT_DISPLAY_WIDTH_DOUBLE 10
+#define DTA_DEFAULT_DISPLAY_WIDTH_STRING 9
 
 #define DTA_FILE_VERSION_MIN     104
 #define DTA_FILE_VERSION_MAX     119
@@ -398,22 +399,35 @@ static readstat_error_t dta_emit_fmtlist(readstat_writer_t *writer, dta_ctx_t *c
             strncpy(&ctx->fmtlist[ctx->fmtlist_entry_len*i],
                     r_variable->format, ctx->fmtlist_entry_len);
         } else {
-            char *format_spec = "9s";
-            if (r_variable->type == READSTAT_TYPE_INT8) {
-                format_spec = DTA_DEFAULT_FORMAT_BYTE;
-            } else if (r_variable->type == READSTAT_TYPE_INT16) {
-                format_spec = DTA_DEFAULT_FORMAT_INT16;
-            } else if (r_variable->type == READSTAT_TYPE_INT32) {
-                format_spec = DTA_DEFAULT_FORMAT_INT32;
-            } else if (r_variable->type == READSTAT_TYPE_FLOAT) {
-                format_spec = DTA_DEFAULT_FORMAT_FLOAT;
-            } else if (r_variable->type == READSTAT_TYPE_DOUBLE) {
-                format_spec = DTA_DEFAULT_FORMAT_DOUBLE;
+            char format_letter = 'g';
+            int display_width = r_variable->display_width;
+            if (readstat_type_class(r_variable->type) == READSTAT_TYPE_CLASS_STRING) {
+                format_letter = 's';
+            }
+            if (!display_width) {
+                if (r_variable->type == READSTAT_TYPE_INT8) {
+                    display_width = DTA_DEFAULT_DISPLAY_WIDTH_BYTE;
+                } else if (r_variable->type == READSTAT_TYPE_INT16) {
+                    display_width = DTA_DEFAULT_DISPLAY_WIDTH_INT16;
+                } else if (r_variable->type == READSTAT_TYPE_INT32) {
+                    display_width = DTA_DEFAULT_DISPLAY_WIDTH_INT32;
+                } else if (r_variable->type == READSTAT_TYPE_FLOAT) {
+                    display_width = DTA_DEFAULT_DISPLAY_WIDTH_FLOAT;
+                } else if (r_variable->type == READSTAT_TYPE_DOUBLE) {
+                    display_width = DTA_DEFAULT_DISPLAY_WIDTH_DOUBLE;
+                } else {
+                    display_width = DTA_DEFAULT_DISPLAY_WIDTH_STRING;
+                }
             }
             char format[64];
-            sprintf(format, "%%%s%s", 
-                    r_variable->alignment == READSTAT_ALIGNMENT_LEFT ? "-" : "",
-                    format_spec);
+            if (format_letter == 'g') {
+                sprintf(format, "%%%s%d.0g", r_variable->alignment == READSTAT_ALIGNMENT_LEFT ? "-" : "",
+                        display_width);
+            } else {
+                sprintf(format, "%%%s%ds",
+                        r_variable->alignment == READSTAT_ALIGNMENT_LEFT ? "-" : "",
+                        display_width);
+            }
             strncpy(&ctx->fmtlist[ctx->fmtlist_entry_len*i],
                     format, ctx->fmtlist_entry_len);
         }