+++ NEWS.md
@@ -1,5 +1,8 @@
 # haven (development version)
 
+* Updated to ReadStat 1.1.4. Most importantly this includes support for
+  SAS binary compression.
+
 # haven 2.3.1
 
 * Add missing methods so `median()`, `quantile()` and `summary()` work